+++ changelog.md
@@ -3,6 +3,8 @@
 ## Unreleased
 
 * Fixed invalid `first_name` validation in the account form modal.
+* Shared public forms now don't allow creating new options
+  for single and multiple select fields.
 * Fixed bug where the arrow keys of a selected cell didn't work when they were not
   rendered.
 * Select new view immediately after creation.